post tut 03

+++ 03/misc/add.c
@@ -2,7 +2,7 @@
 #include <stdint.h>
 
 int main(void) {
-    int16_t x = 30000 + 30000;
+    int32_t x = 30000 + 30000;
 		printf("%d\n", x);
 		return 0;
 }
